test: refactor Smith fuzzer to use the common mz_client
Touches #3709
