+++ test/smith/Cargo.toml
@@ -16,6 +16,7 @@ reqwest = { version = "0.10.6", features = ["json"] }
 serde = { version = "1.0", features = ["derive"] }
 serde_json = "1.0.56"
 structopt = "0.3.15"
+test-util = { path = "../test-util" }
 tokio = { version = "0.2.21", features = ["full"] }
 tokio-postgres = "0.5.4"
 url = "2.1.1"
