--- test/smith/resources/views/view2.sql
@@ -1,1 +1,0 @@
-CREATE MATERIALIZED VIEW view2 (a) AS VALUES ('foo'), ('bar');
